+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 ### Added
 1. Added the `peak_statistics.zero_pad` function to fill sparse arrays with delimiting zero values, and a
    `scan_filter.ZeroFiller` (label: `zero_fill`) filter.
+2. Added `ms_peak_picker.scan_averaging.GridAverager` which makes averaging a fixed set of spectra efficiently
 
 ### Changed
 1. `ms_peak_picker.pick_peaks` no longer returns `None` when the input arrays are empty, instead returning